single computation for both sidebar page mentions and graph
