+++ src/site/_data/eleventyComputed.js
@@ -1,0 +1,5 @@
+const { getGraph } = require("../../helpers/linkUtils");
+
+module.exports = {
+    graph: (data) => getGraph(data),
+}
