+++ src/site/graph.njk
@@ -1,5 +1,5 @@
 ---
-permalink: /graph
+permalink: /graph.json
 eleventyExcludeFromCollections: true
 ---
-{{ graphData | safe }}
+{{ graph | jsonify | safe }}
