+++ src/site/notes/notes.11tydata.js
@@ -1,12 +1,10 @@
 require("dotenv").config();
 const settings = require("../../helpers/constants");
-const { getBacklinks } = require("../../helpers/linkUtils");
 
 const allSettings = settings.ALL_NOTE_SETTINGS;
 
 module.exports = {
   eleventyComputed: {
-    backlinks: (data) => getBacklinks(data),
     settings: (data) => {
       const noteSettings = {};
       allSettings.forEach((setting) => {
